+++ src/Simpletools/Db/Cassandra/Type/Map.php
@@ -1,0 +1,99 @@
+<?php
+
+namespace Simpletools\Db\Cassandra\Type;
+
+class Map implements \JsonSerializable
+{
+	protected $_body;
+	protected $_value;
+	protected $_keyType;
+	protected $_valueType;
+
+	//protected $type;
+	//protected $keys;
+	//protected $values;
+
+	public function __construct($body, $keyType = \Cassandra::TYPE_TEXT, $valueType = \Cassandra::TYPE_TEXT)
+	{
+
+		if (is_array($body)) $body = (object)$body;
+
+		if (is_string($body)) $body = json_decode($body);
+
+		if ($body instanceof \Cassandra\Map)
+		{
+			$this->_body = new \stdClass();
+			$this->_keyType = $body->type()->keyType();
+			$this->_valueType = $body->type()->valueType();
+
+			foreach (array_combine($body->keys(), $body->values()) as $key => $v)
+			{
+				if (substr((string)$v, 0, 1) == '{' && ($obj = json_decode($v)))
+					$this->_body->{$key} = $obj;
+				else
+					$this->_body->{$key} = $v;
+			}
+		}
+		else
+		{
+			$this->_body = $body;
+			$this->_keyType = $keyType;
+			$this->_valueType = $valueType;
+		}
+	}
+
+
+	public function value()
+	{
+		$this->_value = new \Cassandra\Map($this->_keyType, $this->_valueType);
+
+		foreach ($this->_body  as $k =>$v)
+		{
+			if(is_object($v) || is_array($v)) $v = json_encode($v);
+			$this->_value->set($k,$v);
+		}
+
+		return $this->_value;
+	}
+
+	public function jsonSerialize() {
+		return $this->_body;
+	}
+
+	public function __toString()
+	{
+		return json_encode($this->_body);
+	}
+
+	public function __isset($name)
+	{
+		return isset($this->_body->{$name});
+	}
+
+	public function __set($name,$value)
+	{
+		if(is_object($this->_body))
+			$this->_body->{$name} = &$value;
+		else
+			$this->_body = $value;
+	}
+
+	public function __unset($name)
+	{
+		unset($this->_body->{$name});
+	}
+
+	public function &__get($name)
+	{
+		if(property_exists($this->_body, $name))
+		{
+			return $this->_body->{$name};
+		}
+		else
+		{
+			return null;
+		}
+	}
+
+
+}